requirements.txt
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.103.1
h11==0.14.0
idna==3.10
numpy==1.24.3
packaging==24.2
Pillow==10.0.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.0
python-multipart==0.0.6
replicate==0.11.0
requests==2.31.0
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.23.2
